Implement q_remove_head() and fix q_free()
Implement q_remove_head() to remove the node at queue head.

In the previous implementation of q_free(), it doesn't free the head
node when queue is empty. This causes memory leak when program wants
to free the whole queue. This bug is fix by not detecting whether
the queue is empty in the first if condition.

After q_remove_head() is implemented and the q_free() bug is fixed,
q_test can pass all the test cases in trace-01-ops.cmd.