Shuffle the nodes in queue (an alternative)
In previous implementation (commit: 6f86936), after a random node
is picked, I traverse over the doubly linked list to find the node.
The time complexity is O(n^2).

In this commit, I implement a three-pass solution.
In pass one, I traverse over the doubly linked list, delete each
node, and put it into an array of pointers. Therefore, I have an
array of pointers which referece the nodes in queue.
In pass two, run Fisher-Yates algorithm to shuffle the array of
pointers.
In pass three, insert all the nodes into queue.
In this alternative, the time complexity is O(n).